Remove indirection layer, use modules
Remove unused RuntimeDevice, RuntimeQueue, etc.
Introduce Runtime and Compiler modules
Simplify launch code
HostCall: Add default timeout
signals: Add default timeout
