+++ src/compiler_utils.jl
@@ -28,23 +28,6 @@ const jltypes = Dict{Symbol,Type}(
     :double => Float64
 )
 
-# which Julia types map to a given ROC lib type
-const fntypes = Dict{Type,Symbol}(
-    Nothing => :void,
-    Bool    => :bool,
-    Int8    => :i8,
-    Int16   => :i16,
-    Int32   => :i32,
-    Int64   => :i64,
-    UInt8   => :u8,
-    UInt16  => :u16,
-    UInt32  => :u32,
-    UInt64  => :u64,
-    Float16 => :f16,
-    Float32 => :f32,
-    Float64 => :f64
-)
-
 # Decode an expression of the form:
 #
 #    function(arg::arg_type, arg::arg_type, ... arg::arg_type)::return_type
@@ -342,45 +325,3 @@ llvmsize(ty::LLVM.StructType) = ispacked(ty) ? sum(llvmsize(elem) for elem in el
 llvmsize(ty::LLVM.PointerType) = div(Sys.WORD_SIZE, 8)
 llvmsize(ty::LLVM.VectorType) = size(ty)
 llvmsize(ty) = error("Unknown size for type: $ty, typeof: $(typeof(ty))")
-
-@generated function string_length(ex::Union{Ptr,LLVMPtr})
-    Context() do ctx
-        T_ex = convert(LLVMType, ex; ctx)
-        T_ex_ptr = LLVM.PointerType(T_ex)
-        T_i8_ptr = LLVM.PointerType(LLVM.Int8Type(ctx))
-        T_i64 = LLVM.Int64Type(ctx)
-        llvm_f, _ = create_function(T_i64, [T_ex])
-        mod = LLVM.parent(llvm_f)
-        Builder(ctx) do builder
-            entry = BasicBlock(llvm_f, "entry"; ctx)
-            check = BasicBlock(llvm_f, "check"; ctx)
-            done = BasicBlock(llvm_f, "done"; ctx)
-
-            position!(builder, entry)
-            init_offset = ConstantInt(0; ctx)
-            input_ptr = if T_ex isa LLVM.PointerType
-                parameters(llvm_f)[1]
-            else
-                inttoptr!(builder, parameters(llvm_f)[1], T_ex_ptr)
-            end
-            if LLVM.addrspace(llvmtype(input_ptr)) != LLVM.addrspace(T_ex_ptr)
-                input_ptr = addrspacecast!(builder, input_ptr, T_ex_ptr)
-            end
-            input_ptr = bitcast!(builder, input_ptr, T_i8_ptr)
-            br!(builder, check)
-
-            position!(builder, check)
-            offset = phi!(builder, T_i64)
-            next_offset = add!(builder, offset, ConstantInt(1; ctx))
-            append!(LLVM.incoming(offset), [(init_offset, entry), (next_offset, check)])
-            ptr = gep!(builder, input_ptr, [offset])
-            value = load!(builder, ptr)
-            cond = icmp!(builder, LLVM.API.LLVMIntEQ, value, ConstantInt(0x0; ctx))
-            br!(builder, cond, done, check)
-
-            position!(builder, done)
-            ret!(builder, offset)
-        end
-        call_function(llvm_f, Csize_t, Tuple{ex}, :ex)
-    end
-end
