+++ src/device/gcn/helpers.jl
@@ -61,6 +61,23 @@ GCNIntrinsic(jlname, rocname=jlname; isbroken=false, isinverted=false,
              inp_args=(), out_arg=(), roclib=:ocml, suffix=fntypes[first(inp_args)], tobool=false) =
     GCNIntrinsic(jlname, rocname, isbroken, isinverted, inp_args, out_arg, roclib, suffix, tobool)
 
+# which Julia types map to a given ROC lib type
+const fntypes = Dict{Type,Symbol}(
+    Nothing => :void,
+    Bool    => :bool,
+    Int8    => :i8,
+    Int16   => :i16,
+    Int32   => :i32,
+    Int64   => :i64,
+    UInt8   => :u8,
+    UInt16  => :u16,
+    UInt32  => :u32,
+    UInt64  => :u64,
+    Float16 => :f16,
+    Float32 => :f32,
+    Float64 => :f64
+)
+
 import ExprTools
 
 macro device_override(ex)
