+++ src/kernel.jl
@@ -18,7 +18,6 @@ function HSAKernelInstance(agent::HSAAgent, exe::HSAExecutable, symbol::String, 
                            (HSA.Executable, HSA.Agent, HSA.ExecutableSymbol, Ptr{Cvoid}))
     exec_symbol = Ref{HSA.ExecutableSymbol}()
     exec_ptr = Base.unsafe_convert(Ref{Cvoid}, exec_symbol)
-    @debug "Agent symbols"
     HSA.executable_iterate_agent_symbols(exe.executable[], agent.agent,
                                          agent_func, exec_ptr) |> check
 
@@ -88,110 +87,47 @@ function HSAKernelInstance(agent::HSAAgent, exe::HSAExecutable, symbol::String, 
     kernel = HSAKernelInstance(agent, exe, symbol, args, kernel_object[],
                                kernarg_segment_size[], group_segment_size[],
                                private_segment_size[], kernarg_address[])
-    hsaref!()
+    AMDGPU.hsaref!()
     finalizer(kernel) do kernel
         HSA.memory_free(kernel.kernarg_address) |> check
-        hsaunref!()
+        AMDGPU.hsaunref!()
     end
     return kernel
 end
 
-struct HSASignalSet{T}
-    signals::Vector{HSASignal}
-end
-HSASignalSet{T}() where T = HSASignalSet{T}(HSASignal[])
-Base.wait(ss::HSASignalSet{HSA.BarrierAndPacket}) = wait.(ss.signals)
-function Base.wait(ss::HSASignalSet{HSA.BarrierOrPacket})
-    #= FIXME
-    # We need to hack around the fact that barrier OR packets don't handle more
-    # than 5 dependencies. We could implement the waiting in software, and emit
-    # a barrier that waits on a signal tied to that waiter.
-    =#
-    @warn "Waiting on OR barriers waits on all partitioned barriers to complete" maxlog=1
-    wait.(ss.signals)
+## Kernel allocations
+
+struct KernelMetadata
+    kern::UInt64
+    buf::Mem.Buffer
 end
 
-barrier_and!(signals::Vector) = barrier_and!(default_queue().queue, signals)
-barrier_or!(signals::Vector) = barrier_or!(default_queue().queue, signals)
-barrier_and!(queue::HSAQueue, signals::Vector{HSA.Signal}) = barrier_and!(queue, map(HSASignal, signals))
-barrier_or!(queue::HSAQueue, signals::Vector{HSA.Signal}) = barrier_or!(queue, map(HSASignal, signals))
-barrier_and!(queue::HSAQueue, signals::Vector{HSASignal}) =
-    barrier!(HSA.BarrierAndPacket, queue, signals)
-barrier_or!(queue::HSAQueue, signals::Vector{HSASignal}) =
-    barrier!(HSA.BarrierOrPacket, queue, signals)
-function barrier!(T, queue::HSAQueue, signals::Vector{HSASignal})
-    outset = HSASignalSet{T}()
-    if !isempty(signals)
-        for signal_set in Iterators.partition(signals, 5)
-            comp_signal = HSASignal()
-            _launch!(T, queue, comp_signal) do _packet
-                @set! _packet.dep_signal = ntuple(i->length(signal_set)>=i ? signal_set[i].signal[] : HSA.Signal(0), 5)
-                _packet
-            end
-            push!(outset.signals, comp_signal)
-        end
-    end
-    return outset
+## Kernel module and function
+
+export ROCModule, ROCFunction
+
+const MAX_EXCEPTIONS = 256
+const EXE_TO_MODULE_MAP = IdDict{Any,WeakRef}()
+mutable struct ROCModule{E}
+    exe::HSAExecutable{E}
+    metadata::Vector{KernelMetadata}
+    exceptions::Mem.Buffer
 end
-
-# TODO Docstring
-function launch!(queue::HSAQueue, kernel::HSAKernelInstance, signal::HSASignal;
-                 workgroup_size=nothing, grid_size=nothing)
-    @assert workgroup_size !== nothing "Must specify workgroup_size kwarg"
-    @assert grid_size !== nothing "Must specify grid_size kwarg"
-
-    _launch!(HSA.KernelDispatchPacket, queue, signal) do _packet
-        @set! _packet.setup = 3 << Int(HSA.KERNEL_DISPATCH_PACKET_SETUP_DIMENSIONS)
-        @set! _packet.workgroup_size_x = workgroup_size[1]
-        @set! _packet.workgroup_size_y = workgroup_size[2]
-        @set! _packet.workgroup_size_z = workgroup_size[3]
-        @set! _packet.grid_size_x = grid_size[1]
-        @set! _packet.grid_size_y = grid_size[2]
-        @set! _packet.grid_size_z = grid_size[3]
-        @set! _packet.completion_signal = signal.signal[]
-        @set! _packet.kernel_object = kernel.kernel_object
-        @set! _packet.kernarg_address = kernel.kernarg_address
-        @set! _packet.private_segment_size = kernel.private_segment_size
-        @set! _packet.group_segment_size = kernel.group_segment_size
-        _packet
+function ROCModule(exe)
+    agent = exe.agent
+    metadata = KernelMetadata[]
+    exceptions = Mem.alloc(agent, sizeof(AMDGPU.ExceptionEntry)*MAX_EXCEPTIONS; coherent=true)
+    mod = ROCModule(exe, metadata, exceptions)
+    EXE_TO_MODULE_MAP[exe] = WeakRef(mod)
+    AMDGPU.hsaref!()
+    finalizer(mod) do x
+        # FIXME: Free all metadata
+        Mem.free(mod.exceptions)
+        delete!(EXE_TO_MODULE_MAP, exe)
+        AMDGPU.hsaunref!()
     end
 end
-
-function _launch!(f, T, queue::HSAQueue, signal::HSASignal)
-    # Obtain the current queue write index and queue size
-    _queue = unsafe_load(queue.queue)
-    queue_size = _queue.size
-    write_index = HSA.queue_add_write_index_scacq_screl(queue.queue, UInt64(1))
-
-    # Yield until queue has space
-    while write_index - HSA.queue_load_read_index_scacquire(queue.queue) >= queue_size
-        yield()
-    end
-
-    # TODO: Make this less ugly
-    dispatch_packet = Ref{T}()
-    ccall(:memset, Cvoid,
-          (Ptr{Cvoid}, Cint, Csize_t),
-          dispatch_packet, 0, sizeof(T))
-    dispatch_packet[] = f(dispatch_packet[])
-
-    queueMask = UInt32(queue_size - 1)
-    baseaddr_ptr = Ptr{HSA.KernelDispatchPacket}(_queue.base_address)
-    baseaddr_ptr = baseaddr_ptr + sizeof(HSA.KernelDispatchPacket) * (write_index & queueMask)
-    dispatch_packet_ptr = convert(Ptr{HSA.KernelDispatchPacket}, Base.unsafe_convert(Ptr{T}, dispatch_packet))
-    unsafe_copyto!(baseaddr_ptr, dispatch_packet_ptr, 1)
-
-    packetheadertype(::Type{HSA.KernelDispatchPacket}) = HSA.PACKET_TYPE_KERNEL_DISPATCH
-    packetheadertype(::Type{HSA.BarrierAndPacket}) = HSA.PACKET_TYPE_BARRIER_AND
-    packetheadertype(::Type{HSA.BarrierOrPacket}) = HSA.PACKET_TYPE_BARRIER_OR
-
-    # Create and atomically store the header
-    header = Ref{UInt16}(0)
-    header[] |= UInt16(HSA.FENCE_SCOPE_SYSTEM) << UInt16(HSA.PACKET_HEADER_SCACQUIRE_FENCE_SCOPE)
-    header[] |= UInt16(HSA.FENCE_SCOPE_SYSTEM) << UInt16(HSA.PACKET_HEADER_SCRELEASE_FENCE_SCOPE)
-    header[] |= UInt16(packetheadertype(T))    << UInt16(HSA.PACKET_HEADER_TYPE)
-    atomic_store_n!(Base.unsafe_convert(Ptr{UInt16}, baseaddr_ptr), header[])
-
-    # Ring the doorbell to dispatch the kernel
-    HSA.signal_store_relaxed(_queue.doorbell_signal, Int64(write_index))
+mutable struct ROCFunction
+    mod::ROCModule
+    entry::String
 end
