+++ src/mapreduce.jl
@@ -131,12 +131,13 @@ function GPUArrays.mapreducedim!(f::F, op::OP, R::ROCArray{T},
 
     # group size is restricted by local memory
     agent = R.buf.agent
-    pools = filter(pool->pool_segment(pool) == HSA.AMD_SEGMENT_GROUP, memory_pools(agent))
+    pools = filter(pool->Runtime.pool_segment(pool) == HSA.AMD_SEGMENT_GROUP,
+                   Runtime.memory_pools(agent))
     max_items = if !isempty(pools)
         pool = first(pools)
-        max_lmem_elements = pool_size(pool) ÷ sizeof(T)
-        isa = first(isas(agent))
-        Base.min(max_group_size(isa), compute_items(max_lmem_elements ÷ 2))
+        max_lmem_elements = Runtime.pool_size(pool) ÷ sizeof(T)
+        isa = first(Runtime.isas(agent))
+        Base.min(Runtime.max_group_size(isa), compute_items(max_lmem_elements ÷ 2))
     else
         @warn "No group segment detected for agent $agent; assuming 64 elements\nThis message will not be shown again" maxlog=1
         64
