+++ src/reflection.jl
@@ -15,10 +15,10 @@ for method in (:code_typed, :code_warntype, :code_llvm, :code_native)
                          kernel::Bool=false, device=default_device(), kwargs...)
             source = FunctionSpec(func, Base.to_tuple_type(types), kernel)
             isa = default_isa(device)
-            arch = architecture(isa)
-            feat = features(isa)
+            arch = Runtime.architecture(isa)
+            feat = Runtime.features(isa)
             target = GCNCompilerTarget(; dev_isa=arch,features=feat)
-            params = ROCCompilerParams(device, NamedTuple())
+            params = Compiler.ROCCompilerParams(device, NamedTuple())
             job = CompilerJob(target, source, params)
             GPUCompiler.$method($(args...); kwargs...)
         end
