+++ src/sync.jl
@@ -1,3 +1,5 @@
+import ..AMDGPU: hip_configured
+
 "Tracks HSA signals and HIP streams to sync against."
 struct SyncState
     signals::Vector{HSAStatusSignal}
@@ -15,14 +17,14 @@ function wait!(ss::SyncState)
     foreach(wait, ss.signals)
     empty!(ss.signals)
     @static if hip_configured
-        foreach(HIP.hipStreamSynchronize, ss.streams)
+        foreach(AMDGPU.HIP.hipStreamSynchronize, ss.streams)
         empty!(ss.streams)
-        HIP.hipStreamSynchronize(C_NULL) # FIXME: This shouldn't be necessary
+        AMDGPU.HIP.hipStreamSynchronize(C_NULL) # FIXME: This shouldn't be necessary
     end
     nothing
 end
 mark!(ss::SyncState, signal::HSAStatusSignal) = push!(ss.signals, signal)
 mark!(ss::SyncState, stream::Ptr{Cvoid}) = push!(ss.streams, stream)
 
-wait!(x) = adapt(WaitAdaptor(), x)
-mark!(x, s) = adapt(MarkAdaptor(s), x)
+wait!(x) = Adapt.adapt(WaitAdaptor(), x)
+mark!(x, s) = Adapt.adapt(MarkAdaptor(s), x)
