+++ test/hsa/agent.jl
@@ -1,27 +1,27 @@
 @testset "Agent" begin
     @testset "Default selection" begin
-        agent = get_default_agent()
+        agent = AMDGPU.default_device()
         @test agent !== nothing
         @test AMDGPU.device_type(agent) == :gpu
 
-        agent_name = AMDGPU.get_name(agent)
+        agent_name = Runtime.name(agent)
         @test length(agent_name) > 0
         @test !occursin('\0', agent_name)
 
-        if length(AMDGPU.get_agents(:gpu)) > 1
+        if length(AMDGPU.devices()) > 1
             @testset "Multi-GPU" begin
-                init_agent = AMDGPU.get_default_agent()
-                init_dev = AMDGPU.device()
+                init_agent = AMDGPU.default_device()
+                init_dev = AMDGPU.default_device_id()
                 @test init_dev == 1
-                AMDGPU.device!(2)
-                @test AMDGPU.device() == 2
-                @test AMDGPU.get_default_agent() != init_agent
-                AMDGPU.device!(1)
-                @test AMDGPU.device() == 1
-                @test AMDGPU.get_default_agent() == init_agent
+                AMDGPU.default_device_id!(2)
+                @test AMDGPU.default_device_id() == 2
+                @test AMDGPU.default_device() != init_agent
+                AMDGPU.default_device_id!(1)
+                @test AMDGPU.default_device_id() == 1
+                @test AMDGPU.default_device() == init_agent
 
-                @test_throws BoundsError AMDGPU.device!(0)
-                @test_throws BoundsError AMDGPU.device!(length(AMDGPU.get_agents())+1)
+                @test_throws BoundsError AMDGPU.default_device_id!(0)
+                @test_throws BoundsError AMDGPU.default_device_id!(length(AMDGPU.devices())+1)
             end
         else
             @test_skip "Multi-GPU"
@@ -29,8 +29,9 @@
     end
 
     @testset "ISAs" begin
-        agent = get_default_agent()
-        agent_isa = get_first_isa_string(agent)
+        agent = AMDGPU.default_device()
+        agent_isa = string(AMDGPU.default_isa_architecture(agent))
         @test length(agent_isa) > 0
+        @test occursin("gfx", agent_isa)
     end
 end
